[all] Build with -Wshorten-64-to-32

+++ platform/default/headless_view.cpp
@@ -178,7 +178,7 @@ PremultipliedImage HeadlessView::readStillImage() {
     PremultipliedImage image { w, h };
     MBGL_CHECK_ERROR(glReadPixels(0, 0, w, h, GL_RGBA, GL_UNSIGNED_BYTE, image.data.get()));
 
-    const int stride = image.stride();
+    const size_t stride = image.stride();
     auto tmp = std::make_unique<uint8_t[]>(stride);
     uint8_t* rgba = image.data.get();
     for (int i = 0, j = h - 1; i < j; i++, j--) {
